Minor game balancing
The player's automatic fire rate is capped to a set limit, where it makes the game more enjoyable to play, but not easier. This is to add a skill gap between those who purlely use automatic fire and those who use a combination of both tap fire and auto fire, since tap fire has a faster ROF than automatic.